Initial commit for the code structure
Add the initial structure for the system. The
first sketch of the module structure along with
the potential functions with a few comments to
help build the understanding of how this system
is going to look like.

Modules, functions and dependencies are likely to
see drastic changes, but this as this is the first
commit, I think it's alright for it to be just a
sketch.