Catch invalid parameter counts before functions
If user has inputted invalid number of parameters
for a command, the command should fail. In use,
this means that the number of parameters is
checked, verified against allowed number of
parameters for the command in question and
finally decided upon.

If parameters are okay, execute the command as
per usual. If parameters aren't okay, inform
the user about this mistake, print command
usage and return to main menu without executing
the command.