Add authenticative login function
Even when user is logged in, some comamnds
require that the user authenticates themselves
again in an attempt to verify that the user
is indeed who they should be. The most intuitive
way to go about this is to ask the user to log in
again before executing the command.

This is not the most secure or user friendly way
of going about this, but it is the base
implementation I'm going with for now. This can
be switched to something else later on relatively
easily as it's modular. It can also be used for
multiple commands as it is a general
implementation that should work for every
command.