Change how accounts directory string is accessed
As this implementation is already very platform
dependent, it's a good idea to bundle up its
usage so that it can be changed later on more
easily to support other platforms or to allow
changing it mid-run.

Instead of forming the path separately each time
it is needed, a constant could be used so that
the path is only formed once.