Add a way to access main account's services' names
As the user's input specifies the name of the
service, it becomes necessary to address the
services by their names. As the main account
data structure holds a list of service accounts,
they cannot directly be addressed via their names.

A separate function that returns a list of service
names is useful for checking whether a service
exists by addressing it using its name.