Add training command
Login training for the service accounts. This is
the core of the software's use cases. This'll
likely see some changes if new features are to
be added as this is a core functionality for the
users.

For now, it'll just suffice that there's an
infinite loop that asks the user for login
credentials over and over again. Feedback is
given as to whether the user entered correct
username and password or not.