Add main account removal command
This is likely a seldom needed feature, yet
none-the-less important. As main accounts can
be created, they should be removable as well.
No need to go deleting files manually as now
the account can be deleted via the main menu
as well.

The service accounts related to the main account
are deleted as well.