Add service account password change command
The passwords should be changeable. To aid user
with password changing, a strength estimator is
also implemented.

Password length limits of 12-128 were added to
avoid problems caused by overflows and to direct
user's password selection into reasonable and
"secure enough" territory in terms of length.

The strength estimator is still in its infancy as
it only considers bruteforcing. Even as it is, it
could need a little bit of adjusting, but seeing
as this is the initial commit for it, adjustments
can be made later on.