Add main account password change command
Same as with service accounts, main account
password should also be changeable. However,
main account password is stored in a different
manner as it doesn't need to be reversed. Thus
it should be salted and hashed instead of
encrypted.

Implementation for the salting and hashing is
not considered in this commit, but shall be
added later on. This is just the base on top
of which the actual security shall be built.