Changes all around paving way for registration
I noted that there were a few problems with the
design I had in mind, and in order to welcome
main account registration feature into the
program, a lot of changes had to be made. This
would mean that the commit for that would
become way too packed. I split it into smaller
commits.

This one does some refactoring.
- Remove the unused functions that were added as
part of (or around the time of) the initial commit
- Enforce alphabetical order of functions within
modules
- MainAccount salt is optional & defaults to None
- Move password strength display into its own
function
- Move password length validity check to interface

Now the last one here may seem absurd. Why should
functionality be moved to the interface? Well,
as it turned out, registration caused a few
problems with password & username validity
checking as that would require quite a lot of
bouncing in and out from the interface. I
decided that for the time being it's best to add
this validity checking to the interface. This is
a bad choice from the design point of view, but
a mandatory one for the time being. The program
should be gotten to the point where it functions
and I currently don't have the time to redesign
this part to work in a way that complements the
architectural design.

It could be said that this is a bandaid fix that
works for now but must truly be dealt with later
on.

Depency added: interface --uses--> data_formats