Modify the way account info storage works
I noticed that there was a major problem with the
design I had in mind for the account storage. If
the account usernames were to be encrypted or
hashed and used as filenames that way, there'd be
no way to determine the account usernames in that
are already in use as the account passwords would
be needed to decrypt the account usernames either
from the filename or from within the files.

This would mean that there could be different
accounts with the same usernames. As there's
the theoretical possibility that 2 accounts
could have the same password, it wouldn't be
a great scenario as users could log in using
other people's accounts.

To solve this problem, I decided that the main
account names shall not be encrypted nor hashed.
The main accounts can therefore be more easily
distinguished from the service accounts and thus
be used to limit users from creating accounts with
equal names. This doesn't stop users from creating
the account files manually, however.

As the encrypted account username within the file
became useless for main accounts, I decided to
remove that as well. There's no need to make main
accounts and service accounts indistinguishable
anymore so now they can also be separated by the
number of lines within the account files as well.

This is used to speed up the account reading time
as after checking out the length of the content of
the file it can be said which account it is.

This does leave the accounts vulnerable to
manipulation as users can now change the number of
content rows within the files to change which
accounts the program detects to be main, service or
corrupted.

Salt reset was also added to password changing
function within data_formats. This ensures that
when account is stored on disk and a salt is
missing, a new one shall be generated for it.
Salts are not recycled.