Add main account registration support
It would be very silly not to include this one.
Main accounts need to be created at some point
and this feature allows that they're created in
a proper manner.

Account username must be one that is not already
in use, username & password lengths must be
appropriate and the account informations must be
stored on disk in a secure way using the storage
module's implementations.

The secure way of storing and loading in the
storage module's functions hasn't been implemented
yet, but room has been left that this can be
achieved without changing the other modules too
much.