Differentiate main & srv accountfiles by extension
As main and service accountfiles are relatively
easily distinguishable by human eye either way
(as service accounts' filenames shall be encrypted
and the main account' filenames are plaintext),
there's no point in hiding which ones are which.

Therefore the programmatical distinguishing can be
extended into using different file extensions for
main and service accountfiles. It's a little gain
for seemingly no cost.

.account for main accounts
.service for service accounts