Change logging to INFO & improve CLI
As service account encryption will bring a lot of
debug messages, changing the logging to INFO and
up reduces the clutter.

Some UI prints were a little bit cluttered as well
so adding some newlines improved the readability
quite a bit in my opinion.

There was a print in storage module that needn't
be there. Prints shouldn't be used outside of
interface module anyway so this had to be removed.