Add service account encryption & decryption
Service accounts' credentials should also be
stored securely and as the data must be reversable
encryption should be used.

Some changes were made to the encoding and
decoding used in the data storage as the names
of the files will also be encrypted. Therefore
there are restrictions on what characters the
filenames cannot contain (such as "/").

Service account updating hasn't been considered
yet. This means that if main account password is
changed, the service account credentials are lost
as they cannot be decrypted anymore until the main
password is changed back. The service account pw
changes aren't saved yet either.

Added dependency: PyCrypto
- For AES encryption