Add support for encrypted srv account files
As service account's password can be changed, it
must be saved on disk as well. Previously when
the file was in plaintext, the file could be
updated directly. Now that the filename is
encrypted and an unique IV is used for each time
a file is stored, a new file must be created
instead. This means that the older one must be
deleted and a new one shall be created as a
replacement.

Another scenario is that main account's password
gets changed. As service account's filename and
encryption depends on the key derived from the
main account's password, the files must be updated
every time main account's password is changed.
This means that the files are first deleted and
then replaced by new ones.

To make the the file management easier, the
encrypted filename should be stored in the service
account data_format.

During service account removal, user should
authenticate before it's verified whether the
requested account exists.

User should also be notified when the account
removal fails rather than being displayed that the
account removal was successful.